Added domain file and granularity to the deploy process
The CNAME file has been checked in to the public folder, so it gets automatically copied to the GitHub Pages branch. In addition, the formerly called `deploy` script has been renamed to `build-deploy` to show its double commitment, and a new simpler `deploy` script is now in charge to call the `gh-pages` routine only.